refactor: modularize video api client
