+++ src/app/api/search/route.test.ts
@@ -1,8 +1,23 @@
 import { NextRequest } from 'next/server';
+import { describe, it, expect, beforeEach, vi } from 'vitest';
+
+vi.mock('@/features/videos/api/nocodb', () => ({
+  fetchAllVideos: vi.fn(),
+  videoListItemSchema: {},
+}));
+
+import { fetchAllVideos } from '@/features/videos/api/nocodb';
 import { GET } from './route';
 
+const mockedFetchAllVideos = vi.mocked(fetchAllVideos);
+
 describe('/api/search', () => {
+  beforeEach(() => {
+    mockedFetchAllVideos.mockReset();
+  });
+
   it('should return empty results for empty query', async () => {
+    mockedFetchAllVideos.mockResolvedValue([] as never);
     const request = new NextRequest('http://localhost:3000/api/search?q=');
     const response = await GET(request);
     const data = await response.json();
@@ -14,8 +29,7 @@ describe('/api/search', () => {
   });
 
   it('should handle search requests correctly', async () => {
-    // Mock the fetchAllVideos function
-    const mockVideos = [
+    mockedFetchAllVideos.mockResolvedValue([
       {
         Id: 1,
         Title: 'Test Video',
@@ -26,20 +40,9 @@ describe('/api/search', () => {
         Persons: [{ Title: 'John Doe' }],
         CreatedAt: new Date('2023-01-01'),
       },
-      {
-        Id: 2,
-        Title: 'Another Video',
-        Description: 'Another test video',
-        Channel: 'Another Channel',
-        VideoID: 'test456',
-        Hashtags: ['another', 'test'],
-        Persons: [{ Title: 'Jane Smith' }],
-        CreatedAt: new Date('2023-01-02'),
-      },
-    ];
+    ] as never);
 
-    // This test would need to mock the database call
-    // For now, we'll just test the API structure
+    // Validate that the endpoint responds with the expected shape.
     const request = new NextRequest('http://localhost:3000/api/search?q=test&categories=title,description');
     const response = await GET(request);
     const data = await response.json();
